+++ src/services/users/index.ts
@@ -1,4 +1,4 @@
-const { User } = require("../../../models");
+const { User } = require("../../models");
 
 exports.getUsers = async () => {
   const res = await User.findAll();
